完成自动匹配计算

+++ momo-run/src/main/java/com/momohelp/calculate/service/ICalculation.java
@@ -8,7 +8,7 @@ import com.momohelp.model.Farm;
 public interface ICalculation {
 
     //奖金基数计算
-	public abstract int base();
+	public abstract double base();
 
 	//买卖盘自动匹配
 	public abstract  boolean automatch();
